refactor: auth 패키지 구조 기능별로 구분되도록 변경
