+++ server/src/main/java/wap/starlist/auth/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package wap.starlist.config.auth;
+package wap.starlist.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
