+++ server/src/main/java/wap/starlist/auth/dto/OAuth2UserPrincipal.java
@@ -1,4 +1,4 @@
-package wap.starlist.config.auth.dto;
+package wap.starlist.auth.dto;
 
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
