+++ server/src/main/java/wap/starlist/auth/dto/UserRole.java
@@ -1,4 +1,4 @@
-package wap.starlist.config.auth.dto;
+package wap.starlist.auth.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
