+++ server/src/main/java/wap/starlist/auth/security/TokenAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package wap.starlist.config.auth;
+package wap.starlist.auth.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -9,10 +9,11 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
+import wap.starlist.auth.JwtTokenProvider;
 
 import java.io.IOException;
 
-import static wap.starlist.config.auth.AuthConstants.*;
+import static wap.starlist.auth.constants.AuthConstants.*;
 
 @Component
 @RequiredArgsConstructor
